Rename Truncated => TruncatedSpan, make available as v-truncated
