+++ src/components/Functional/TruncatedSpan.vue
